==> scripts/perception/2_cleaning.Rmd <==
---
title: "2_cleaning"
author: "Nick Kathios"
date: '2023-01-16'
output: html_document
---
#load packages
```{r}
library(tidyverse)
library(lme4)
library(mediation)
library(lmerTest)
library(performance)
library(sjPlot)
library(boot)
```

#read in data
```{r}
scored_data <- read.csv(file="../../data/pleasing-displeasing_followup/scored_data.csv")
master <- read.csv(file="../../data/pleasing-displeasing_followup/pleasing_displeasing_scored.csv")
```

#clean data
```{r}
#removing Ps w/o ELA measures
scored_data <- scored_data[complete.cases(scored_data[ , 391:418]),]
#removing Ps w/o STAI
scored_data <- scored_data[complete.cases(scored_data[ , 431]),]
```

#cleaning original data before merge
```{r}
master <- master[-1]
names(master)[1] <- "ID"
names(master)[10:21] <- c('A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J','K','L')
names(master)[22:31] <-c('IADS105','IADS110','IADS220','IADS283','IADS351','IADS353',
                         'IADS423','IADS502','IADS717','IADS730')
names(master)[32:39] <- c('Envi1','Envi2','Envi3','Envi4','Envi5','Envi6','Envi7','Envi8')

just_ratings<-master[10:39] -4
cleaned <- cbind(master[1], just_ratings)
cleaned <- cbind(cleaned, master[267])

derived <- cleaned[1:13]
derivedstims <- pivot_longer(derived, A:L, names_to = "stim_name", values_to = "rating" )

IADS <- cleaned[14:23]
IADS <- cbind(cleaned[1], IADS)
IADS <- cbind(cleaned[32], IADS)
IADS_stims <- pivot_longer(IADS, IADS105:IADS730, names_to = "stim_name", values_to = "rating" )

envi <- just_ratings[23:30]
envi <- cbind(cleaned[1], envi)
envi_stims <- pivot_longer(envi, Envi1:Envi8, names_to = "stim_name", values_to = "rating" )
#changing stim names for pleasing/displeasing comparisons
names(derived)[2:13] <- c("Displeasing1", "Displeasing2", "Pleasing3","Displeasing4",
                          "Pleasing5", "Pleasing6", "Displeasing7","Displeasing8",
                          "Pleasing9", "Pleasing10", "Pleasing11", "Displeasing12")

derived_plsdispls <- pivot_longer(derived, !ID, names_to = "soundID", values_to = "rating" )

derived_plsdispls$stim_name <- gsub("[0-9]", "", derived_plsdispls$soundID)

names(IADS)[3:12] <- c("Displeasing1", "Pleasing2", "Pleasing3", "Displeasing4",
                          "Pleasing5", "Pleasing6", "Displeasing7","Displeasing8",
                          "Pleasing9", "Displeasing10")
IADS_plsdispls <- pivot_longer(IADS, !PASS_total:ID, names_to = "soundID", values_to = "rating")

IADS_plsdispls$stim_name <- gsub("[0-9]", "", IADS_plsdispls$soundID)
```

#merge the datasets
```{r}
names(scored_data)[20] <- "ID"

all_scored <- merge(master, scored_data, by.x="ID", by.y="ID")
digital <-merge(derived_plsdispls, all_scored, by.x="ID", by.y="ID")
digital_pleasing <- subset(digital, stim_name=="Pleasing")
digital_displeasing <- subset(digital, stim_name=="Displeasing")
IADS<-merge(IADS_plsdispls, all_scored, by.x="ID", by.y="ID")
IADS_pleasing <- subset(IADS, stim_name=="Pleasing")
names(IADS_pleasing)[2] <-"PASS_total"
IADS_displeasing <- subset(IADS, stim_name=="Displeasing")

write.csv(all_scored, "../../data/pleasing-displeasing_followup/perception/all_scored.csv")
write.csv(digital_pleasing, "../../data/pleasing-displeasing_followup/perception/digital_pleasing.csv")
write.csv(digital_displeasing, "../../data/pleasing-displeasing_followup/perception/digital_displeasing.csv")
write.csv(IADS_pleasing, "../../data/pleasing-displeasing_followup/perception/IADS_pleasing.csv")
write.csv(IADS_displeasing, "../../data/pleasing-displeasing_followup/perception/IADS_displeasing.csv")
```
