==> scripts/perception/4_followupanalysis.Rmd <==
---
title: "4_followupanalysis"
author: "Nick Kathios"
date: '2023-02-25'
output: html_document
---
#load packages
```{r}
library(tidyverse)
library(lme4)
library(lmerTest)
```

#read data in
```{r}
followup <- read.csv(file='../../data/SLfollowupscored.csv')
dg_conditions <- read.csv(file='../../data/SL_followup/dg_conditions.csv')
ogdata<-read.csv(file='../../data/pleasing-displeasing_followup/scored_data.csv')

#removing individuals from the pleasing/displeasing study
keep <- setdiff(followup$ID,ogdata$Q158)
followup_norepeats <- followup[followup$ID %in% keep, ] 
```

#things to check for: probably don't want to include those from the original pls/displs study
#if longform CHAOS doesn't work out, could try with shortform scores or questions that closely index auditory commotion
#or, could try dimensionality reduction with the ratings?

#cleaning
```{r}
dgratings<- followup_norepeats %>%
  dplyr::select("A_dgroughness_1":"L_dgpleasantness_1")

roughness_index <- seq(1,24,by=2)
pleasure_index <- seq(2,24, by=2)

roughness <- dgratings[roughness_index] %>%
  cbind(followup_norepeats[2]) %>%
  pivot_longer(!ID, names_to = "stim", values_to = "rating")

x=1:nrow(roughness)
for (i in x) {
 roughness$stim[i] <- gsub("_dgroughness_1", "",roughness$stim[i])
}
roughness_master <- merge(dg_conditions, roughness)

roughnessdispleasing_master <- subset(roughness_master, condition=="displeasing")

roughnessdispleasing_master <- merge(roughnessdispleasing_master, followup_norepeats)

roughnesspleasing_master <- subset(roughness_master, condition=="pleasing")

roughnesspleasing_master <- merge(roughnesspleasing_master, followup_norepeats)

pleasantness <- dgratings[pleasure_index] %>%
  cbind(followup_norepeats[2]) %>%
  pivot_longer(!ID, names_to = "stim", values_to = "rating")

x=1:nrow(pleasantness)
for (i in x) {
 pleasantness$stim[i] <- gsub("_dgpleasantness_1", "",pleasantness$stim[i])
}
pleasantness_master <- merge(dg_conditions, pleasantness)

pleasantnessdispleasing_master <- subset(pleasantness_master, condition=="displeasing")

pleasantnessdispleasing_master <- merge(pleasantnessdispleasing_master, followup_norepeats)

pleasantnesspleasing_master <- subset(pleasantness_master, condition=="pleasing")

pleasantnesspleasing_master <- merge(pleasantnesspleasing_master, followup_norepeats)
```

#MET Correlation Plots
```{r}
#ELA Sum Scores
variables = dplyr::select(followup, 
                          MET_melodicscore, MET_rhythmicscore,
                          CHAOS_sum, QUIC_sum, Threat_sum, deprivation_sum,
                          Gold_musicaltraining)
                          
output = corr.test(variables)

corr_matrix <- output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('MET Melodic','MET Rhythmic','CHAOS Sum', 'QUIC Sum', 'Threat Sum', 'Deprivation Sum', 'GoldMSI - Musical Training')
rownames(corr_matrix) <- c('MET Melodic','MET Rhythmic','CHAOS Sum', 'QUIC Sum', 'Threat Sum', 'Deprivation Sum', 'GoldMSI - Musical Training')

corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1))

#ELA Childhood
variables = dplyr::select(followup, 
                          MET_melodicscore, MET_rhythmicscore,
                          CHAOS_child_total, QUIC_child_total, threat_childhood_total, deprivation_child_total,
                          Gold_musicaltraining)
                          
output = corr.test(variables)

corr_matrix <- output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('MET Melodic','MET Rhythmic','CHAOS Childhood', 'QUIC Childhood', 'Threat Childhood', 'Deprivation Childhood', 'GoldMSI - Musical Training')
rownames(corr_matrix) <- c('MET Melodic','MET Rhythmic','CHAOS Childhood', 'QUIC Childhood', 'Threat Childhood', 'Deprivation Childhood', 'GoldMSI - Musical Training')

corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1))

#ELA Adolescence
variables = dplyr::select(followup, 
                          MET_melodicscore, MET_rhythmicscore,
                          CHAOS_Adolescence_total, QUIC_Adolescence_total, 
                          threat_adolescence_total, deprivation_Adolescence_total,
                          Gold_musicaltraining)
                          
output = corr.test(variables)

corr_matrix <- output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('MET Melodic','MET Rhythmic','CHAOS Adolescence', 'QUIC Adolescence', 'Threat Adolescence', 'Deprivation Childhood', 'GoldMSI - Musical Training')
rownames(corr_matrix) <- c('MET Melodic','MET Rhythmic','CHAOS Adolescence', 'QUIC Adolescence', 'Threat Adolescence', 'Deprivation Childhood', 'GoldMSI - Musical Training')

corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1))
```

#MET Models
```{r}
m1 <- lm(MET_melodicscore ~ QUIC_sum + CHAOS_sum + Threat_sum + deprivation_sum + PASS_total + STAI_total + Gold_musicaltraining, data = followup)
summary(m1)

m1a <- lm(MET_melodicscore ~ QUIC_child_total + CHAOS_child_total + threat_childhood_total + deprivation_child_total + PASS_total + STAI_total + Gold_musicaltraining, data = followup)
summary(m1a)

m1b <- lm(MET_melodicscore ~ QUIC_Adolescence_total + CHAOS_Adolescence_total + threat_adolescence_total + deprivation_Adolescence_total + PASS_total + STAI_total + Gold_musicaltraining, data = followup)
summary(m1b)

m2 <- lm(MET_rhythmicscore ~ QUIC_sum + CHAOS_sum + Threat_sum + deprivation_sum + PASS_total + STAI_total + Gold_musicaltraining, data = followup)
summary(m2)

m2a <- lm(MET_rhythmicscore ~ QUIC_child_total + CHAOS_child_total + threat_childhood_total + deprivation_child_total + PASS_total + STAI_total + Gold_musicaltraining, data = followup)
summary(m2a)

m2b <- lm(MET_rhythmicscore ~ QUIC_Adolescence_total + CHAOS_Adolescence_total + threat_adolescence_total + deprivation_Adolescence_total + PASS_total + STAI_total + Gold_musicaltraining, data = followup)
summary(m2b)
```

#Displeasing Models
```{r}
m1 <- lmer(rating ~ QUIC_sum + CHAOS_sum + Threat_sum + deprivation_sum + (1 | ID) + (1 | stim), data = pleasantnessdispleasing_master)
summary(m1)

m1a <- lmer(rating ~ QUIC_child_total + CHAOS_child_total + threat_childhood_total + deprivation_child_total + (1 | ID) + (1 | stim), data = pleasantnessdispleasing_master)
summary(m1a)

m1b <- lmer(rating ~ QUIC_Adolescence_total + CHAOS_Adolescence_total + threat_adolescence_total + deprivation_Adolescence_total + (1 | ID) + (1 | stim), data = pleasantnessdispleasing_master)
summary(m1b)

m2 <- lmer(rating ~ QUIC_sum + CHAOS_sum + Threat_sum + deprivation_sum + (1 | ID) + (1 | stim), data = roughnessdispleasing_master)
summary(m2)

m2a <- lmer(rating ~ QUIC_child_total + CHAOS_child_total + threat_childhood_total + deprivation_child_total + (1 | ID) + (1 | stim), data = roughnessdispleasing_master)
summary(m2a)

m2b <- lmer(rating ~ QUIC_Adolescence_total + CHAOS_Adolescence_total + threat_adolescence_total + deprivation_Adolescence_total + (1 | ID) + (1 | stim), data = roughnessdispleasing_master)
summary(m2b)
```

#Pleasing Models
```{r}
m1 <- lmer(rating ~ QUIC_sum + CHAOS_sum + Threat_sum + deprivation_sum + (1 | ID) + (1 | stim), data = pleasantnesspleasing_master)
summary(m1)

m1a <- lmer(rating ~ QUIC_child_total + CHAOS_child_total + threat_childhood_total + deprivation_child_total + (1 | ID) + (1 | stim), data = pleasantnesspleasing_master)
summary(m1a)

m1b <- lmer(rating ~ QUIC_Adolescence_total + CHAOS_Adolescence_total + threat_adolescence_total + deprivation_Adolescence_total + (1 | ID) + (1 | stim), data = pleasantnesspleasing_master)
summary(m1b)

m2 <- lmer(rating ~ QUIC_sum + CHAOS_sum + Threat_sum + deprivation_sum + (1 | ID) + (1 | stim), data = roughnesspleasing_master)
summary(m2)

m2a <- lmer(rating ~ QUIC_child_total + CHAOS_child_total + threat_childhood_total + deprivation_child_total + (1 | ID) + (1 | stim), data = roughnesspleasing_master)
summary(m2a)

m2b <- lmer(rating ~ QUIC_Adolescence_total + CHAOS_Adolescence_total + threat_adolescence_total + deprivation_Adolescence_total + (1 | ID) + (1 | stim), data = roughnesspleasing_master)
summary(m2b)
```

#Scatterplots
```{r}
#CHAOS Child --> MET melodic
CHAOS_child_METmelodic <- 
  ggplot(followup, aes(x=CHAOS_child_total, y=MET_melodicscore)) +
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("Total Child CHAOS Score") +
  ylab("MET Melodic") +
  theme(axis.title=element_text(size=20,face="bold"))
print(CHAOS_child_METmelodic)
ggsave("CHAOS_child_METmelodic.png", width = 6.5, height = 4)

#CHAOS Child --> MET rhythmic
CHAOS_child_MET_rhythmic <- 
  ggplot(followup, aes(x=CHAOS_child_total, y=MET_rhythmicscore)) +
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("Total Child CHAOS Score") +
  ylab("MET Rhythmic") +
  theme(axis.title=element_text(size=20,face="bold"))
print(CHAOS_child_MET_rhythmic)
ggsave("CHAOS_child_METrhythmic.png", width = 6.5, height = 4)

#Threat Adolescence --> MET melodic
threat_adol_METmelodic <- 
  ggplot(followup, aes(x=threat_adolescence_total, y=MET_melodicscore)) +
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("Total Adolescent Threat Score") +
  ylab("MET Melodic") +
  theme(axis.title=element_text(size=20,face="bold"))
print(threat_adol_METmelodic)
ggsave("threat_adol_METmelodic.png", width = 6.5, height = 4)

#Threat Adolescence --> MET rhythmic
threat_adol_METrhythmic <- 
  ggplot(followup, aes(x=threat_adolescence_total, y=MET_rhythmicscore)) +
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("Total Adolescent Threat Score") +
  ylab("MET Rhythmic") +
  theme(axis.title=element_text(size=20,face="bold"))
print(threat_adol_METrhythmic)
ggsave("threat_adol_METrhythmic.png", width = 6.5, height = 4)

```
