==> scripts/perception/5_scoring_sample3.Rmd <==
---
title: "5_scoring_sample3"
author: "Kelsie Lopez"
date: "2023-07-27"
output: html_document
---

#packages
```{r}
library(tidyverse)
```

#read in .csv
```{r}
survey <- read.csv(file="../../../POS-DEV-Scale/data/POS DEV CFA_May 10, 2023_10.52 2.csv")
df <- read.csv(file="../../../TEAM/data/perception/Perception Covariate Follow-Up_September 8, 2023_16.13.csv")
source("../reverseCode.R")
```

#d + t scoring 
```{r}
### CHILDHOOD ###
#Deprivation Scoring
deprivation_child <- survey %>%
  dplyr::select("DT_child1_1":"DT_child40_1")

cols_reverse <- c("DT_child1_1", "DT_child2_1", "DT_child3_1", "DT_child4_1", "DT_child5_1", "DT_child6_1", "DT_child7_1", "DT_child10_1", "DT_child19_1", "DT_child20_1", "DT_child21_1", "DT_child22_1", "DT_child23_1", "DT_child24_1", "DT_child25_1", "DT_child26_1", "DT_child27_1", "DT_child28_1", "DT_child29_1", "DT_child30_1", "DT_child31_1", "DT_child32_1", "DT_child33_1", "DT_child35_1", "DT_child36_1", "DT_child38_1", "DT_child39_1", "DT_child40_1")

cols_emotionaldeprivation_inhome <- c("DT_child1_1", "DT_child2_1", "DT_child3_1", "DT_child4_1", "DT_child5_1", "DT_child6_1", "DT_child7_1")

cols_physicaldeprivation_inhome <- c("DT_child8_1", "DT_child9_1", "DT_child10_1", "DT_child11_1", "DT_child12_1", "DT_child13_1", "DT_child14_1", "DT_child15_1", "DT_child16_1", "DT_child17_1", "DT_child18_1")

cols_cognitivedeprivation_inhome <- c("DT_child19_1", "DT_child20_1", "DT_child21_1", "DT_child22_1", "DT_child23_1", "DT_child24_1", "DT_child25_1", "DT_child26_1")

cols_emotionaldeprivation_outofhome <- c("DT_child27_1", "DT_child28_1", "DT_child29_1", "DT_child30_1", "DT_child31_1", "DT_child32_1", "DT_child33_1")

cols_physicaldeprivation_outofhome <- c("DT_child34_1")

cols_cognitivedeprivation_outofhome <- c("DT_child35_1", "DT_child36_1", "DT_child37_1", "DT_child38_1", "DT_child39_1", "DT_child40_1")



# Reverse code
deprivation_child[cols_reverse] <- lapply(deprivation_child[cols_reverse], reverseCode, min=1, max=4)

#scoring
survey$deprivation_child_total <- apply(deprivation_child,1, sum)
survey$deprivation_child_emotional_inhome <- apply(deprivation_child[cols_emotionaldeprivation_inhome], 1,sum)
survey$deprivation_child_physical_inhome <- apply(deprivation_child[cols_physicaldeprivation_inhome], 1,sum)
survey$deprivation_child_cognitive_inhome <- apply(deprivation_child[cols_cognitivedeprivation_inhome], 1,sum)
survey$deprivation_child_emotional_outofhome <- apply(deprivation_child[cols_emotionaldeprivation_outofhome], 1,sum)
survey$deprivation_child_physical_outofhome <- apply(deprivation_child[cols_physicaldeprivation_outofhome], 1,sum)
survey$deprivation_child_cognitive_outofhome <- apply(deprivation_child[cols_cognitivedeprivation_outofhome], 1,sum)
hist(survey$deprivation_child_total)

### ADOLESCENCE ###
#Deprivation Scoring
deprivation_Adolescence <- survey %>%
  dplyr::select("DT_adol1_1":"DT_adol38_1")

cols_reverse <- c("DT_adol1_1", "DT_adol2_1", "DT_adol3_1", "DT_adol4_1", "DT_adol5_1", "DT_adol6_1", "DT_adol7_1", "DT_adol8_1", "DT_adol17_1", "DT_adol18_1", "DT_adol19_1", "DT_adol20_1", "DT_adol21_1", "DT_adol22_1", "DT_adol23_1", "DT_adol24_1", "DT_adol25_1", "DT_adol26_1", "DT_adol27_1", "DT_adol28_1", "DT_adol29_1", "DT_adol30_1", "DT_adol31_1", "DT_adol33_1", "DT_adol34_1", "DT_adol36_1", "DT_adol37_1", "DT_adol38_1")

cols_emotionaldeprivation_inhome <- c("DT_adol1_1", "DT_adol2_1", "DT_adol3_1", "DT_adol4_1", "DT_adol5_1", "DT_adol6_1", "DT_adol7_1")

cols_physicaldeprivation_inhome <- c("DT_adol8_1", "DT_adol9_1", "DT_adol10_1", "DT_adol11_1", "DT_adol12_1", "DT_adol13_1", "DT_adol14_1", "DT_adol15_1", "DT_adol16_1")

cols_cognitivedeprivation_inhome <- c("DT_adol17_1", "DT_adol18_1", "DT_adol19_1", "DT_adol20_1", "DT_adol21_1", "DT_adol22_1", "DT_adol23_1", "DT_adol24_1")

cols_emotionaldeprivation_outofhome <- c("DT_adol25_1", "DT_adol26_1", "DT_adol27_1", "DT_adol28_1", "DT_adol29_1", "DT_adol30_1", "DT_adol31_1")

cols_physicaldeprivation_outofhome <- c("DT_adol32_1")

cols_cognitivedeprivation_outofhome <- c("DT_adol33_1", "DT_adol34_1", "DT_adol35_1", "DT_adol36_1", "DT_adol37_1", "DT_adol38_1")

# Reverse code
deprivation_Adolescence[cols_reverse] <- lapply(deprivation_Adolescence[cols_reverse], reverseCode, min=1, max=4)

#scoring
survey$deprivation_Adolescence_total <- apply(deprivation_Adolescence,1, sum)
survey$deprivation_Adolescence_emotional_inhome <- apply(deprivation_Adolescence[cols_emotionaldeprivation_inhome], 1,sum)
survey$deprivation_Adolescence_physical_inhome <- apply(deprivation_Adolescence[cols_physicaldeprivation_inhome], 1,sum)
survey$deprivation_Adolescence_cognitive_inhome <- apply(deprivation_Adolescence[cols_cognitivedeprivation_inhome], 1,sum)
survey$deprivation_Adolescence_emotional_outofhome <- apply(deprivation_Adolescence[cols_emotionaldeprivation_outofhome], 1,sum)
survey$deprivation_Adolescence_physical_outofhome <- apply(deprivation_Adolescence[cols_physicaldeprivation_outofhome], 1,sum)
survey$deprivation_Adolescence_cognitive_outofhome <- apply(deprivation_Adolescence[cols_cognitivedeprivation_outofhome], 1,sum)
hist(survey$deprivation_Adolescence_total)

survey$deprivation_sum <- survey$deprivation_child_total + survey$deprivation_Adolescence_total

### CHILDHOOD ###
#Threat Scoring
threat_child <- survey %>%
  dplyr::select("DT_child41_1": "DT_child106_1", "DT_child115_1":"DT_child118_1") %>%
  dplyr::select(-c("DT_child44", "DT_child46", "DT_child47", "DT_child49", "DT_child50", "DT_child52", "DT_child53", "DT_child55", "DT_child56", "DT_child58", "DT_child59", "DT_child61", "DT_child62", "DT_child64", "DT_child65", "DT_child67", "DT_child68", "DT_child70", "DT_child71", "DT_child73", "DT_child74", "DT_child76", "DT_child77", "DT_child79", "DT_child80", "DT_child83", "DT_child84", "DT_child86", "DT_child87", "DT_child89", "DT_child90", "DT_child92", "DT_child93", "DT_child95", "DT_child96", "DT_child98", "DT_child99", "DT_child101", "DT_child102", "DT_child104", "DT_child105"))

cols_home_threat <- c("DT_child41_1", "DT_child42_1", "DT_child43_1","DT_child45_1", "DT_child48_1", "DT_child51_1", "DT_child54_1","DT_child57_1", "DT_child60_1","DT_child63_1", "DT_child66_1", "DT_child69_1", "DT_child72_1", "DT_child75_1", "DT_child78_1")
cols_neighborhood_threat <- c("DT_child81_1", "DT_child82_1", "DT_child85_1", "DT_child88_1", "DT_child91_1", "DT_child94_1", "DT_child97_1", "DT_child100_1", "DT_child103_1", "DT_child106_1")
cols_identity_threat <- c("DT_child115_1", "DT_child116_1", "DT_child117_1", "DT_child118_1")

survey$threat_childhood_total <- apply(threat_child,1, sum)
survey$threat_childhood_home <- apply(threat_child[cols_home_threat], 1,sum)
survey$threat_childhood_neighborhood <- apply(threat_child[cols_neighborhood_threat], 1,sum) 
survey$threat_childhood_identity <- apply(threat_child[cols_identity_threat], 1,sum)
hist(survey$threat_childhood_total) #fix***

survey$threat_childhood_total

### ADOLESCENCE ###
#Threat Scoring
threat_adolescence <- survey %>%
  dplyr::select("DT_adol39_1": "DT_adol104_1", "DT_adol113_1":"DT_adol116_1") %>%
  dplyr::select(-c("DT_adol42", "DT_adol44", "DT_adol45", "DT_adol47", "DT_adol48", "DT_adol50", "DT_adol51", "DT_adol53", "DT_adol54", "DT_adol56", "DT_adol57", "DT_adol59", "DT_adol60", "DT_adol62", "DT_adol63", "DT_adol65", "DT_adol66", "DT_adol68", "DT_adol69", "DT_adol71", "DT_adol72", "DT_adol74", "DT_adol75", "DT_adol77", "DT_adol78", "DT_adol81", "DT_adol82", "DT_adol84", "DT_adol85", "DT_adol87", "DT_adol88", "DT_adol90", "DT_adol91", "DT_adol93", "DT_adol94", "DT_adol96", "DT_adol97", "DT_adol99", "DT_adol100", "DT_adol102", "DT_adol103"))

cols_home_threat <- c("DT_adol39_1", "DT_adol40_1", "DT_adol41_1", "DT_adol43_1", "DT_adol46_1", "DT_adol49_1", "DT_adol52_1", "DT_adol55_1", "DT_adol58_1", "DT_adol61_1", "DT_adol64_1", "DT_adol67_1", "DT_adol70_1", "DT_adol73_1", "DT_adol76_1")
cols_neighborhood_threat <- c("DT_adol79_1", "DT_adol80_1", "DT_adol83_1", "DT_adol86_1", "DT_adol89_1", "DT_adol92_1", "DT_adol95_1", "DT_adol98_1", "DT_adol101_1", "DT_adol104_1")
cols_identity_threat <- c("DT_adol113_1", "DT_adol114_1", "DT_adol115_1", "DT_adol116_1")

survey$threat_adolescence_total <- apply(threat_adolescence,1, sum)
survey$threat_adolescence_home <- apply(threat_adolescence[cols_home_threat], 1,sum)
survey$threat_adolescence_neighborhood <- apply(threat_adolescence[cols_neighborhood_threat], 1,sum)
survey$threat_adolescence_identity <- apply(threat_adolescence[cols_identity_threat], 1,sum)
hist(survey$threat_adolescence_total) 

ggplot(data=survey, aes(x=threat_childhood_total, y=threat_adolescence_total)) + geom_point()+
  geom_smooth()

survey$Threat_sum <- survey$threat_childhood_total + survey$threat_adolescence_total
```

#CHAOS
```{r}
#child
CHAOS_child <- survey %>%
  dplyr::select("CHAOS_child1_1":"CHAOS_child15_1")

cols_reverse <- c("CHAOS_child1_1", "CHAOS_child2_1", "CHAOS_child4_1", "CHAOS_child7_1", "CHAOS_child12_1", "CHAOS_child14_1", "CHAOS_child15_1")

cols_shortform <- c("CHAOS_child4_1", "CHAOS_child6_1", "CHAOS_child10_1", "CHAOS_child13_1",
                   "CHAOS_child14_1", "CHAOS_child15_1")

# Reverse code
CHAOS_child[cols_reverse] <- lapply(CHAOS_child[cols_reverse], reverseCode, min=1, max=5)

survey$CHAOS_child <- apply(CHAOS_child,1,sum)
survey$CHAOS_child_shortform <- apply(CHAOS_child[cols_shortform],1, sum)
hist(survey$CHAOS_child)

#adolescent
CHAOS_adol <- survey %>%
  dplyr::select("CHAOS_adol1_1":"CHAOS_adol15_1")

cols_reverse <- c("CHAOS_adol1_1", "CHAOS_adol2_1", "CHAOS_adol4_1", "CHAOS_adol7_1", "CHAOS_adol12_1", "CHAOS_adol14_1", "CHAOS_adol15_1")

cols_shortform <- c("CHAOS_adol4_1", "CHAOS_adol6_1", "CHAOS_adol10_1", "CHAOS_adol13_1",
                   "CHAOS_adol14_1", "CHAOS_adol15_1")

# Reverse code
CHAOS_adol[cols_reverse] <- lapply(CHAOS_adol[cols_reverse], reverseCode, min=1, max=5)

survey$CHAOS_adol <- apply(CHAOS_adol,1,sum)
survey$CHAOS_adol_shortform <- apply(CHAOS_adol[cols_shortform],1, sum)
hist(survey$CHAOS_adol)

survey$CHAOS_sum <- survey$CHAOS_adol + survey$CHAOS_child
survey$CHAOS_sum_shortform <- survey$CHAOS_adol_shortform + survey$CHAOS_child_shortform
```

#QUIC
```{r}
#child
QUIC_child <- survey %>%
  dplyr::select("QUIC_child1":"QUIC_child38")


QUIC_child[QUIC_child == "1"] <- as.numeric(1)
QUIC_child[QUIC_child == "4"] <- as.numeric(0)
QUIC_child[1:38] <- as.numeric(unlist(QUIC_child[1:38]))

cols_reverse <- c("QUIC_child1", "QUIC_child3", "QUIC_child4", "QUIC_child5", "QUIC_child6", "QUIC_child7", "QUIC_child8", "QUIC_child9", "QUIC_child10", "QUIC_child14", "QUIC_child15", "QUIC_child17", "QUIC_child28", "QUIC_child36")

cols_parental_monitoring <- c("QUIC_child1", "QUIC_child3", "QUIC_child4", "QUIC_child5", "QUIC_child6", "QUIC_child7", "QUIC_child9", "QUIC_child10", "QUIC_child14")

cols_parental_predictability <- c("QUIC_child2", "QUIC_child8", "QUIC_child11", "QUIC_child12", "QUIC_child15", "QUIC_child16", "QUIC_child17", "QUIC_child31", "QUIC_child32", "QUIC_child33", "QUIC_child34", "QUIC_child35")

cols_parental_enivironment <- c("QUIC_child18", "QUIC_child19", "QUIC_child21", "QUIC_child22", "QUIC_child28", "QUIC_child29", "QUIC_child30")

cols_parental_enivironment <- c("QUIC_child13", "QUIC_child20", "QUIC_child26", "QUIC_child27", "QUIC_child36", "QUIC_child37", "QUIC_child38")

cols_safety_security <- c("QUIC_child23", "QUIC_child24", "QUIC_child25")

# Reverse code
QUIC_child[cols_reverse] <- lapply(QUIC_child[cols_reverse], reverseCode, min=0, max=1)

survey$QUIC_child_total <- apply(QUIC_child,1, sum)
survey$QUIC_child_parental_monitoring <- apply(QUIC_child[cols_parental_monitoring], 1,sum)
survey$QUIC_child_parental_predictability<- apply(QUIC_child[cols_parental_predictability], 1,sum)
survey$QUIC_child_parental_enivironment <- apply(QUIC_child[cols_parental_enivironment], 1,sum)
survey$QUIC_child_safety_security <- apply(QUIC_child[cols_safety_security], 1,sum)
hist(survey$QUIC_child_total)

#adolescent
QUIC_adol <- survey %>%
  dplyr::select("QUIC_adol1":"QUIC_adol36")

QUIC_adol[QUIC_adol == "1"] <- as.numeric(1)
QUIC_adol[QUIC_adol == "4"] <- as.numeric(0)
QUIC_adol[1:36] <- as.numeric(unlist(QUIC_adol[1:36]))

cols_reverse <- c("QUIC_adol1", "QUIC_adol3", "QUIC_adol4", "QUIC_adol5", "QUIC_adol6", "QUIC_adol7", "QUIC_adol8", "QUIC_adol12", "QUIC_adol13", "QUIC_adol15", "QUIC_adol26", "QUIC_adol34")

cols_parental_monitoring <- c("QUIC_adol1", "QUIC_adol3", "QUIC_adol4", "QUIC_adol5", "QUIC_adol7", "QUIC_adol8", "QUIC_adol12")

cols_parental_predictability <- c("QUIC_adol2", "QUIC_adol6", "QUIC_adol9", "QUIC_adol10", "QUIC_adol13", "QUIC_adol14", "QUIC_adol15", "QUIC_adol29", "QUIC_adol30", "QUIC_adol31", "QUIC_adol32", "QUIC_adol33")

cols_parental_enivironment <- c("QUIC_adol16", "QUIC_adol17", "QUIC_adol19", "QUIC_adol20", "QUIC_adol26", "QUIC_adol27", "QUIC_adol28")

cols_physical_enivironment <- c("QUIC_adol11", "QUIC_adol18", "QUIC_adol24", "QUIC_adol25", "QUIC_adol34", "QUIC_adol35", "QUIC_adol36")

cols_safety_security <- c("QUIC_adol21", "QUIC_adol22", "QUIC_adol23")

# Reverse code
QUIC_adol[cols_reverse] <- lapply(QUIC_adol[cols_reverse], reverseCode, min=0, max=1)

survey$QUIC_adol_total <- apply(QUIC_adol,1, sum)
survey$QUIC_adol_parental_monitoring <- apply(QUIC_adol[cols_parental_monitoring], 1,sum)
survey$QUIC_adol_parental_predictability<- apply(QUIC_adol[cols_parental_predictability], 1,sum)
survey$QUIC_adol_parental_enivironment <- apply(QUIC_adol[cols_parental_enivironment], 1,sum)
survey$QUIC_adol_physical_enivironment <- apply(QUIC_adol[cols_physical_enivironment], 1,sum)
survey$QUIC_adol_safety_security <- apply(QUIC_adol[cols_safety_security], 1,sum)
survey$QUIC_sum <- survey$QUIC_child_total + survey$QUIC_adol_total

hist(survey$QUIC_adol_total)
```

#STAI
```{r}
survey$STAI_10_1
STAI <- survey %>%
  dplyr::select("STAI_1_1":"STAI_20_1")


cols_reverse <- c("STAI_1_1","STAI_3_1","STAI_6_1","STAI_7_1","STAI_10_1","STAI_13_1","STAI_14_1","STAI_16_1","STAI_19_1")
STAI[cols_reverse] <- lapply(STAI[cols_reverse], reverseCode, min=1, max=4)


#if it is a simple subscore (i.e. no reverse coding or anything)
survey$STAI_total <- apply(STAI,1,sum)
hist(survey$STAI_total)
```

#PAS Scoring
```{r}
PAS <- df %>%
  dplyr::select("PAS_1":"PAS_61")

#delete attention-checking question
PAS$catch <- NULL

PAS[PAS == 1] <- 1
# Replace 2s with 1s
PAS[PAS == 2] <- 0

# PAS columns to reverse code and sub-score (for these questions, the correct answer (ie the more anhedonic answer) is false.) --> look at PAS scoring guidelines

cols_reverse <- c("PAS_1","PAS_2","PAS_3","PAS_4","PAS_7","PAS_18","PAS_20","PAS_23","PAS_24","PAS_29","PAS_30","PAS_33","PAS_34","PAS_35","PAS_36","PAS_37","PAS_38","PAS_39","PAS_41","PAS_44","PAS_45","PAS_46","PAS_48","PAS_52","PAS_53","PAS_57","PAS_58","PAS_59","PAS_60","PAS_61")

# Reverse code

PAS[cols_reverse] <- lapply(PAS[cols_reverse], reverseCode, min=0, max=1)

# Add sub-scores and total to main df

df$PAS_total <- apply(PAS,1, sum) # total PAS scores

```

#GoldMSI Scoring
```{r}
Gold <- df %>%
  dplyr::select("GMSI_1_1":"GMSI_38")

#delete atttention checking question
Gold$Qcheck <- NULL


# GoldMSI columns to reverse code and sub-score (for these questions, the correct answer (ie the more positive answer) is a higher number. Negative questions are reverse coded.) 

cols_reverse <- c("GMSI_21_1", "GMSI_11_1", "GMSI_13_1", "GMSI_23_1","GMSI_14_1","GMSI_27_1" , "GMSI_17_1" , "GMSI_25_1" , 
"GMSI_9_1")

cols_activeengagement <- c("GMSI_1_1" , "GMSI_3_1" , "GMSI_8_1" , "GMSI_15_1" , "GMSI_21_1" , "GMSI_24_1", "GMSI_28_1", "GMSI_34", "GMSI_38")

cols_perceptualabilities <- c("GMSI_5_1" , "GMSI_6_1" , "GMSI_11_1" , "GMSI_12_1" , "GMSI_13_1" , "GMSI_18_1" , "GMSI_22_1" , "GMSI_23_1" , 
"GMSI_26_1")

cols_musicaltraining <- c("GMSI_14_1", "GMSI_27_1", "GMSI_32", "GMSI_33", "GMSI_35", "GMSI_36", "GMSI_37")


cols_singingabilities <- c("GMSI_4_1" , "GMSI_7_1" , "GMSI_10_1" , "GMSI_17_1" , "GMSI_25_1" , "GMSI_29_1" , "GMSI_30_1")


cols_emotion <- c("GMSI_2_1" , "GMSI_9_1" , "GMSI_16_1" , "GMSI_19_1" , "GMSI_20_1" , "GMSI_31_1")

cols_musicalsophistication <- c("GMSI_1_1" , "GMSI_3_1" , "GMSI_4_1" , "GMSI_7_1" , "GMSI_10_1" , "GMSI_12_1" , "GMSI_14_1" , "GMSI_15_1" , "GMSI_17_1" , "GMSI_19_1" , "GMSI_23_1" , "GMSI_24_1" , "GMSI_25_1" , "GMSI_27_1" , "GMSI_29_1" , "GMSI_32" , "GMSI_33" , "GMSI_37")

# Reverse code

Gold[cols_reverse] <- lapply(Gold[cols_reverse], reverseCode, min=1, max=7)

# Add sub-scores and total to main df

df$Gold_total <- apply(Gold,1, sum)/38 # total GoldMSI score
df$Gold_activeengagement <- apply(Gold[cols_activeengagement], 1,sum)/9
df$Gold_perceptualabilities <- apply(Gold[cols_perceptualabilities],1, sum)/9
df$Gold_musicaltraining <-apply(Gold[cols_musicaltraining],1, sum)/7
df$Gold_singingabilities <- apply(Gold[cols_singingabilities],1, sum)/7
df$Gold_emotion <- apply(Gold[cols_emotion],1, sum)/6
df$Gold_musicalsophistication <- apply(Gold[cols_musicalsophistication],1, sum)/18
```

#POSDEV
```{r}
#childhood caregiver
cols_safetyandsupport_child_caregiver <- c("POS_child19_1","POS_child35_1", "POS_child37_1")
cols_opportunity_child_caregiver <- c("POS_child22_1", "POS_child24_1","POS_child25_1","POS_child26_1","POS_child29_1")
cols_predictability_child_caregiver <- c("POS_child2_1","POS_child13_1","POS_child5_1") #ADD 6 BACK IN

survey$safetyandsupport_child_caregiver <- apply(survey[cols_safetyandsupport_child_caregiver], 1,sum)
survey$opportunity_child_caregiver <- apply(survey[cols_opportunity_child_caregiver],1, sum)
survey$predictability_child_caregiver <-apply(survey[cols_predictability_child_caregiver],1, sum)

survey$general_child_caregiver <- survey$safetyandsupport_child_caregiver + survey$opportunity_child_caregiver + survey$predictability_child_caregiver

ggplot(data=survey, aes(x=general_child_caregiver))+
  geom_histogram() +
  ylab("Count") +
  xlab("Total Child Caregiver Score") +
  theme_classic()

#adolescent caregiver
cols_safetyandsupport_adol_caregiver <- c("POS_adol15_1","POS_adol16_1","POS_adol19_1","POS_adol33_1","POS_adol34_1","POS_adol35_1" ,"POS_adol36_1","POS_adol37_1")
cols_opportunity_adol_caregiver <- c("POS_adol22_1","POS_adol23_1","POS_adol24_1","POS_adol25_1","POS_adol26_1","POS_adol27_1","POS_adol29_1")
cols_predictability_adol_caregiver <- c("POS_adol2_1","POS_adol3_1","POS_adol4_1","POS_adol5_1","POS_adol6_1")

survey$safetyandsupport_adol_caregiver <- apply(survey[cols_safetyandsupport_adol_caregiver], 1,sum)
survey$opportunity_adol_caregiver <- apply(survey[cols_opportunity_adol_caregiver],1, sum)
survey$predictability_adol_caregiver <-apply(survey[cols_predictability_adol_caregiver],1, sum)
survey$general_adol_caregiver <- survey$safetyandsupport_adol_caregiver + survey$opportunity_adol_caregiver + survey$predictability_adol_caregiver

ggplot(data=survey, aes(x=general_adol_caregiver))+
  geom_histogram() +
  ylab("Count") +
  xlab("Total Adolescent Caregiver Score") +
  theme_classic()

#childhood friend
cols_safetyandsupport_child_friend <- c("POS_child72_1","POS_child79_1","POS_child80_1","POS_child84_1","POS_child85_1")
cols_opportunity_child_friend <- c("POS_child76_1","POS_child77_1","POS_child78_1","POS_child82_1","POS_child83_1")

survey$safetyandsupport_child_friend <- apply(survey[cols_safetyandsupport_child_friend], 1,sum)
survey$opportunity_child_friend <- apply(survey[cols_opportunity_child_friend],1, sum)
survey$general_child_friend <- survey$safetyandsupport_child_friend + survey$opportunity_child_friend

ggplot(data=survey, aes(x=general_child_friend))+
  geom_histogram(bins=20) +
  ylab("Count") +
  xlab("Total Child Friend Score") +
  theme_classic()

#adolescent friend
cols_safetyandsupport_adol_friend <- c("POS_adol72_1","POS_adol75_1","POS_adol79_1","POS_adol80_1","POS_adol81_1","POS_adol84_1","POS_adol85_1")
cols_opportunity_adol_friend <- c("POS_adol76_1","POS_adol78_1","POS_adol82_1","POS_adol83_1")
cols_predictability_adol_friend <- c("POS_adol67_1","POS_adol69_1","POS_adol70_1","POS_adol71_1")

survey$safetyandsupport_adol_friend <- apply(survey[cols_safetyandsupport_adol_friend], 1,sum)
survey$opportunity_adol_friend <- apply(survey[cols_opportunity_adol_friend],1, sum)
survey$predictability_adol_friend <-apply(survey[cols_predictability_adol_friend],1, sum)
survey$general_adol_friend <- survey$safetyandsupport_adol_friend + survey$opportunity_adol_friend + survey$predictability_adol_friend

ggplot(data=survey, aes(x=predictability_adol_friend))+
  geom_histogram(bins=15) +
  ylab("Count") +
  xlab("Adolescent Friend Predictability Score") +
  theme_classic()

ggplot(data=survey, aes(x=opportunity_adol_friend))+
  geom_histogram(bins=15) +
  ylab("Count") +
  xlab("Adolescent Friend Opportunity Score") +
  theme_classic()

ggplot(data=survey, aes(x=safetyandsupport_adol_friend))+
  geom_histogram(bins=15) +
  ylab("Count") +
  xlab("Adolescent Friend Safety Score") +
  theme_classic()


ggplot(data=survey, aes(x=general_adol_friend))+
  geom_histogram(bins=15) +
  ylab("Count") +
  xlab("Total Adolescent Friend Score") +
  theme_classic()

#childhood adult role model
cols_safetyandsupport_child_rolemodel <- c("POS_child8_RM_1","POS_child20_RM_1","POS_child22_RM_1","POS_child26_RM_1")
cols_opportunity_child_rolemodel <- c("POS_child10_RM_1","POS_child12_RM_1","POS_child24_RM_1")

survey$safetyandsupport_child_rolemodel <- apply(survey[cols_safetyandsupport_child_rolemodel], 1,sum)
survey$opportunity_child_rolemodel <- apply(survey[cols_opportunity_child_rolemodel],1, sum)
survey$general_child_rolemodel <- survey$safetyandsupport_child_rolemodel + survey$opportunity_child_rolemodel

ggplot(data=survey, aes(x=general_child_rolemodel))+
  geom_histogram(bins=15) +
  ylab("Count") +
  xlab("Total Child Role Model Score") +
  theme_classic()


#adolescent role model
cols_safetyandsupport_adol_rolemodel <- c("POS_adol20_RM_1","POS_adol22_RM_1","POS_adol26_RM_1")
cols_opportunity_adol_rolemodel <- c("POS_adol10_RM_1","POS_adol12_RM_1","POS_adol24_RM_1")

survey$safetyandsupport_adol_rolemodel <- apply(survey[cols_safetyandsupport_adol_rolemodel], 1,sum)
survey$opportunity_adol_rolemodel <- apply(survey[cols_opportunity_adol_rolemodel],1, sum)
survey$general_adol_rolemodel <- survey$safetyandsupport_adol_rolemodel + survey$opportunity_adol_rolemodel

ggplot(data=survey, aes(x=general_adol_rolemodel))+
  geom_histogram(bins=7) +
  ylab("Count") +
  xlab("Total Adolescent Role Model Score") +
  theme_classic()


## sum scores ##
#removing NAs for 0s for *only* posdev total scores 
survey[672:692][is.na(survey[672:692])] <- 0

survey$POSDEVchild_total <- (survey$general_child_caregiver + survey$general_child_friend + survey$general_child_rolemodel)

ggplot(data=survey, aes(x=POSDEVchild_total))+
  geom_histogram() +
  ylab("Count") +
  xlab("Total Child Score") +
  theme_classic()

survey$POSDEVadol_total <- (survey$general_adol_caregiver + survey$general_adol_friend + survey$general_adol_rolemodel)

ggplot(data=survey, aes(x=POSDEVadol_total))+
  geom_histogram() +
  ylab("Count") +
  xlab("Total Adolescent Score") +
  theme_classic()

survey$POSDEVpredictability_total <- (survey$predictability_child_caregiver + survey$predictability_adol_caregiver + survey$predictability_adol_friend)

ggplot(data=survey, aes(x=POSDEVpredictability_total))+
  geom_histogram() +
  ylab("Count") +
  xlab("Total Predictability Score") +
  theme_classic()

survey$POSDEVsafety_total <- (survey$safetyandsupport_child_caregiver + survey$safetyandsupport_adol_caregiver + survey$safetyandsupport_child_friend + survey$safetyandsupport_adol_friend + survey$safetyandsupport_child_rolemodel + survey$safetyandsupport_adol_rolemodel)

ggplot(data=survey, aes(x=POSDEVsafety_total))+
  geom_histogram() +
  ylab("Count") +
  xlab("Total Safety Score") +
  theme_classic()

survey$POSDEVopportunity_total <- (survey$opportunity_child_caregiver + survey$opportunity_adol_caregiver + survey$opportunity_child_friend + survey$opportunity_adol_friend + survey$opportunity_child_rolemodel + survey$opportunity_adol_rolemodel)

ggplot(data=survey, aes(x=POSDEVopportunity_total))+
  geom_histogram() +
  ylab("Count") +
  xlab("Total Opportunity Score") +
  theme_classic()

survey[672:692][survey[672:692] == 0] <- NA
```

#MBEA1 
```{r}
MBEA1 <- survey %>%
  dplyr::select(X1_mbea1_ans:X61_mbea1_ans)
names(MBEA1)[1:31] <- c('MBEA1_1','MBEA1_2','MBEA1_3','MBEA1_4','MBEA1_5','MBEA1_6','MBEA1_7','MBEA1_8',
                      'MBEA1_9','MBEA1_10','MBEA1_11','MBEA1_12','MBEA1_13','MBEA1_14','MBEA1_15','MBEA1_16','MBEA1_17',
                      'MBEA1_18','MBEA1_19','MBEA1_20','MBEA1_21','MBEA1_22','MBEA1_23','MBEA1_24','MBEA1_25','MBEA1_26',
                      'MBEA1_27','MBEA1_28','MBEA1_29','MBEA1_30','MBEA1_31')
#scoring cols w/ same as answer
MBEA1$MBEA1_1 = ifelse(MBEA1$MBEA1_1 == '1', 1, 0)
MBEA1$MBEA1_3 = ifelse(MBEA1$MBEA1_3 == '1', 1, 0)
MBEA1$MBEA1_4 = ifelse(MBEA1$MBEA1_4 == '1', 1, 0)
MBEA1$MBEA1_5 = ifelse(MBEA1$MBEA1_5 == '1', 1, 0)
MBEA1$MBEA1_10 = ifelse(MBEA1$MBEA1_10 == '1', 1, 0)
MBEA1$MBEA1_13 = ifelse(MBEA1$MBEA1_13 == '1', 1, 0)
MBEA1$MBEA1_15 = ifelse(MBEA1$MBEA1_15 == '1', 1, 0)
MBEA1$MBEA1_16 = ifelse(MBEA1$MBEA1_16 == '1', 1, 0)
MBEA1$MBEA1_18 = ifelse(MBEA1$MBEA1_18 == '1', 1, 0)
MBEA1$MBEA1_20 = ifelse(MBEA1$MBEA1_20 == '1', 1, 0)
MBEA1$MBEA1_21 = ifelse(MBEA1$MBEA1_21 == '1', 1, 0)
MBEA1$MBEA1_22 = ifelse(MBEA1$MBEA1_22 == '1', 1, 0)
MBEA1$MBEA1_25 = ifelse(MBEA1$MBEA1_25 == '1', 1, 0)
MBEA1$MBEA1_30 = ifelse(MBEA1$MBEA1_30 == '1', 1, 0)
MBEA1$MBEA1_31 = ifelse(MBEA1$MBEA1_31 == '1', 1, 0)
#scoring cols w/ different as answer
MBEA1$MBEA1_2 = ifelse(MBEA1$MBEA1_2 == '2', 1, 0)
MBEA1$MBEA1_6 = ifelse(MBEA1$MBEA1_6 == '2', 1, 0)
MBEA1$MBEA1_8 = ifelse(MBEA1$MBEA1_8 == '2', 1, 0)
MBEA1$MBEA1_9 = ifelse(MBEA1$MBEA1_9 == '2', 1, 0)
MBEA1$MBEA1_11 = ifelse(MBEA1$MBEA1_11 == '2', 1, 0)
MBEA1$MBEA1_12 = ifelse(MBEA1$MBEA1_12 == '2', 1, 0)
MBEA1$MBEA1_14 = ifelse(MBEA1$MBEA1_14 == '2', 1, 0)
MBEA1$MBEA1_17 = ifelse(MBEA1$MBEA1_17 == '2', 1, 0)
MBEA1$MBEA1_19 = ifelse(MBEA1$MBEA1_19 == '2', 1, 0)
MBEA1$MBEA1_23 = ifelse(MBEA1$MBEA1_23 == '2', 1, 0)
MBEA1$MBEA1_24 = ifelse(MBEA1$MBEA1_24 == '2', 1, 0)
MBEA1$MBEA1_26 = ifelse(MBEA1$MBEA1_26 == '2', 1, 0)
MBEA1$MBEA1_27 = ifelse(MBEA1$MBEA1_27 == '2', 1, 0)
MBEA1$MBEA1_28 = ifelse(MBEA1$MBEA1_28 == '2', 1, 0)
MBEA1$MBEA1_29 = ifelse(MBEA1$MBEA1_29 == '2', 1, 0)

#take out catch trial
MBEA1$MBEA1_7 = NULL

survey$MBEA1_total <- apply(MBEA1,1, sum)
```

#MBEA2
```{r}
MBEA2 <- survey %>%
  dplyr::select(X1_Q1888:X61_Q1888)
names(MBEA2)[1:31] <- c('MBEA2_1','MBEA2_2','MBEA2_3','MBEA2_4','MBEA2_5','MBEA2_6','MBEA2_7','MBEA2_8',
                      'MBEA2_9','MBEA2_10','MBEA2_11','MBEA2_12','MBEA2_13','MBEA2_14','MBEA2_15','MBEA2_16','MBEA2_17',
                      'MBEA2_18','MBEA2_19','MBEA2_20','MBEA2_21','MBEA2_22','MBEA2_23','MBEA2_24','MBEA2_25','MBEA2_26',
                      'MBEA2_27','MBEA2_28','MBEA2_29','MBEA2_30','MBEA2_31')
#scoring cols w/ same as answer
MBEA2$MBEA2_2 = ifelse(MBEA2$MBEA2_2 == '1', 1, 0)
MBEA2$MBEA2_3 = ifelse(MBEA2$MBEA2_3 == '1', 1, 0)
MBEA2$MBEA2_4 = ifelse(MBEA2$MBEA2_4 == '1', 1, 0)
MBEA2$MBEA2_5 = ifelse(MBEA2$MBEA2_5 == '1', 1, 0)
MBEA2$MBEA2_8 = ifelse(MBEA2$MBEA2_8 == '1', 1, 0)
MBEA2$MBEA2_10 = ifelse(MBEA2$MBEA2_10 == '1', 1, 0)
MBEA2$MBEA2_11 = ifelse(MBEA2$MBEA2_11 == '1', 1, 0)
MBEA2$MBEA2_13 = ifelse(MBEA2$MBEA2_13 == '1', 1, 0)
MBEA2$MBEA2_15 = ifelse(MBEA2$MBEA2_15 == '1', 1, 0)
MBEA2$MBEA2_16 = ifelse(MBEA2$MBEA2_16 == '1', 1, 0)
MBEA2$MBEA2_17 = ifelse(MBEA2$MBEA2_17 == '1', 1, 0)
MBEA2$MBEA2_19 = ifelse(MBEA2$MBEA2_19 == '1', 1, 0)
MBEA2$MBEA2_22 = ifelse(MBEA2$MBEA2_22 == '1', 1, 0)
MBEA2$MBEA2_27 = ifelse(MBEA2$MBEA2_27 == '1', 1, 0)
MBEA2$MBEA2_29 = ifelse(MBEA2$MBEA2_29 == '1', 1, 0)
#scoring cols w/ different as answer
MBEA2$MBEA2_1 = ifelse(MBEA2$MBEA2_1 == '2', 1, 0)
MBEA2$MBEA2_6 = ifelse(MBEA2$MBEA2_6 == '2', 1, 0)
MBEA2$MBEA2_7 = ifelse(MBEA2$MBEA2_7 == '2', 1, 0)
MBEA2$MBEA2_9 = ifelse(MBEA2$MBEA2_9 == '2', 1, 0)
MBEA2$MBEA2_12 = ifelse(MBEA2$MBEA2_12 == '2', 1, 0)
MBEA2$MBEA2_14 = ifelse(MBEA2$MBEA2_14 == '2', 1, 0)
MBEA2$MBEA2_18 = ifelse(MBEA2$MBEA2_18 == '2', 1, 0)
MBEA2$MBEA2_20 = ifelse(MBEA2$MBEA2_20 == '2', 1, 0)
MBEA2$MBEA2_23 = ifelse(MBEA2$MBEA2_23 == '2', 1, 0)
MBEA2$MBEA2_24 = ifelse(MBEA2$MBEA2_24 == '2', 1, 0)
MBEA2$MBEA2_25 = ifelse(MBEA2$MBEA2_25 == '2', 1, 0)
MBEA2$MBEA2_26 = ifelse(MBEA2$MBEA2_26 == '2', 1, 0)
MBEA2$MBEA2_28 = ifelse(MBEA2$MBEA2_28 == '2', 1, 0)
MBEA2$MBEA2_30 = ifelse(MBEA2$MBEA2_30 == '2', 1, 0)
MBEA2$MBEA2_31 = ifelse(MBEA2$MBEA2_31 == '2', 1, 0)

#take out catch trial
MBEA2$MBEA2_21 = NULL

survey$MBEA2_total <- apply(MBEA2,1, sum)
```

#MBEA4
```{r}
MBEA4 <- survey %>%
  dplyr::select(X1_Q1882:X61_Q1882)
names(MBEA4)[1:31] <- c('MBEA4_1','MBEA4_2','MBEA4_3','MBEA4_4','MBEA4_5','MBEA4_6','MBEA4_7','MBEA4_8',
                      'MBEA4_9','MBEA4_10','MBEA4_11','MBEA4_12','MBEA4_13','MBEA4_14','MBEA4_15','MBEA4_16','MBEA4_17',
                      'MBEA4_18','MBEA4_19','MBEA4_20','MBEA4_21','MBEA4_22','MBEA4_23','MBEA4_24','MBEA4_25','MBEA4_26',
                      'MBEA4_27','MBEA4_28','MBEA4_29','MBEA4_30','MBEA4_31')
#scoring cols w/ same as answer
MBEA4$MBEA4_1 = ifelse(MBEA4$MBEA4_1 == '1', 1, 0)
MBEA4$MBEA4_4 = ifelse(MBEA4$MBEA4_4 == '1', 1, 0)
MBEA4$MBEA4_5 = ifelse(MBEA4$MBEA4_5 == '1', 1, 0)
MBEA4$MBEA4_8 = ifelse(MBEA4$MBEA4_8 == '1', 1, 0)
MBEA4$MBEA4_11 = ifelse(MBEA4$MBEA4_11 == '1', 1, 0)
MBEA4$MBEA4_15 = ifelse(MBEA4$MBEA4_15 == '1', 1, 0)
MBEA4$MBEA4_16 = ifelse(MBEA4$MBEA4_16 == '1', 1, 0)
MBEA4$MBEA4_20 = ifelse(MBEA4$MBEA4_20 == '1', 1, 0)
MBEA4$MBEA4_21 = ifelse(MBEA4$MBEA4_21 == '1', 1, 0)
MBEA4$MBEA4_23 = ifelse(MBEA4$MBEA4_23 == '1', 1, 0)
MBEA4$MBEA4_25 = ifelse(MBEA4$MBEA4_25 == '1', 1, 0)
MBEA4$MBEA4_26 = ifelse(MBEA4$MBEA4_26 == '1', 1, 0)
MBEA4$MBEA4_27 = ifelse(MBEA4$MBEA4_27 == '1', 1, 0)
MBEA4$MBEA4_29 = ifelse(MBEA4$MBEA4_29 == '1', 1, 0)
MBEA4$MBEA4_30 = ifelse(MBEA4$MBEA4_30 == '1', 1, 0)
#scoring cols w/ different as answer
MBEA4$MBEA4_2 = ifelse(MBEA4$MBEA4_2 == '2', 1, 0)
MBEA4$MBEA4_3 = ifelse(MBEA4$MBEA4_3 == '2', 1, 0)
MBEA4$MBEA4_6 = ifelse(MBEA4$MBEA4_6 == '2', 1, 0)
MBEA4$MBEA4_7 = ifelse(MBEA4$MBEA4_7 == '2', 1, 0)
MBEA4$MBEA4_9 = ifelse(MBEA4$MBEA4_9 == '2', 1, 0)
MBEA4$MBEA4_10 = ifelse(MBEA4$MBEA4_10 == '2', 1, 0)
MBEA4$MBEA4_12 = ifelse(MBEA4$MBEA4_12 == '2', 1, 0)
MBEA4$MBEA4_13 = ifelse(MBEA4$MBEA4_13 == '2', 1, 0)
MBEA4$MBEA4_17 = ifelse(MBEA4$MBEA4_17 == '2', 1, 0)
MBEA4$MBEA4_18 = ifelse(MBEA4$MBEA4_18 == '2', 1, 0)
MBEA4$MBEA4_19 = ifelse(MBEA4$MBEA4_19 == '2', 1, 0)
MBEA4$MBEA4_22 = ifelse(MBEA4$MBEA4_22 == '2', 1, 0)
MBEA4$MBEA4_24 = ifelse(MBEA4$MBEA4_24 == '2', 1, 0)
MBEA4$MBEA4_28 = ifelse(MBEA4$MBEA4_28 == '2', 1, 0)
MBEA4$MBEA4_31 = ifelse(MBEA4$MBEA4_31 == '2', 1, 0)

#take out catch trial
MBEA4$MBEA4_14 = NULL

survey$MBEA4_total <- apply(MBEA4,1, sum)
```

#saving output
```{r}
all <- merge(x= survey, y= df, by= 'PROLIFIC_PID')
write.csv(all, file="../../../TEAM/data/perception/survey_scored_sample3.csv")
```
